requirements.txt
# FastAPI и зависимости
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# База данных
sqlalchemy==2.0.25
asyncpg==0.29.0

# Безопасность
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# Для работы с датами
python-dateutil==2.8.2

# Опционально (для production)
gunicorn==21.2.0
yookassa==3.7.1
YooMoney==0.1.2
alembic==1.15.2
aiofiles==24.1.0
aiogram==3.22.0
asyncio==4.0.0
psycopg2-binary==2.9.10
prometheus-fastapi-instrumentator==6.1.0
